refactor(frontend): extract shared grapheme date and field error helpers and unauthenticated redirect
