+++ frontend/src/components/cardgroups/cardgroup-list-item.tsx
@@ -1,14 +1,11 @@
 import Link from "next/link";
+import { formatMediumDate } from "@/lib/format";
 
 export type CardgroupListItemProps = {
   id: string;
   name: string;
   updatedAt: string;
 };
-
-function formatDate(iso: string): string {
-  return new Intl.DateTimeFormat("en-US", { dateStyle: "medium" }).format(new Date(iso));
-}
 
 export function CardgroupListItem({ id, name, updatedAt }: CardgroupListItemProps) {
   return (
@@ -18,7 +15,7 @@ export function CardgroupListItem({ id, name, updatedAt }: CardgroupListItemProp
         className="flex flex-col gap-1 rounded-lg border border-border p-4 hover:bg-accent transition-colors"
       >
         <span className="font-medium text-foreground">{name}</span>
-        <span className="text-sm text-muted-foreground">Updated {formatDate(updatedAt)}</span>
+        <span className="text-sm text-muted-foreground">Updated {formatMediumDate(updatedAt)}</span>
       </Link>
     </li>
   );
