+++ frontend/src/schemas/cardgroup.ts
@@ -1,13 +1,7 @@
 import { z } from "zod";
+import { graphemeCount } from "./grapheme";
 
 // Mirrors NewCardgroupInput / UpdateCardgroupInput in schema/schema.graphql.
-// UAX #29 grapheme cluster counting keeps FE and BE length rules in sync.
-const segmenter = new Intl.Segmenter(undefined, { granularity: "grapheme" });
-
-function graphemeCount(s: string): number {
-  return Array.from(segmenter.segment(s)).length;
-}
-
 const cardgroupNameSchema = z
   .string()
   .trim()
